+++ docs/source/conf.py
@@ -50,6 +50,7 @@ extensions = [
     "sphinx.ext.intersphinx",  # links to numpy, scipy ... docs
     "sphinx.ext.coverage",
     "sphinx.ext.extlinks",  # define roles for links
+    "sphinx.ext.viewcode",
 ]
 
 # If you need extensions of a certain version or higher, list them here.
@@ -77,6 +78,23 @@ html_theme_options = {
     "light_logo": "logos/dark/logo_primary.svg",
     "dark_logo": "logos/light/logo_primary.svg",
     "sidebar_hide_name": True,
+    "footer_icons": [
+        {
+            "name": "GitHub",
+            "url": "https://github.com/dylanljones/pyrekordbox",
+            "html": """
+                <svg stroke="currentColor" fill="currentColor" stroke-width="0" viewBox="0 0 16 16">
+                    <path fill-rule="evenodd" d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.013 8.013 0 0 0 16 8c0-4.42-3.58-8-8-8z"></path>
+                </svg>
+            """,
+            "class": "",
+        },
+    ],
+    "light_css_variables": {
+        "color-brand-primary": "red",
+        "color-brand-content": "#CC3333",
+        "color-admonition-background": "orange",
+    },
 }
 
 # html_title = f"{release} documentation"
@@ -144,20 +162,22 @@ intersphinx_mapping = {
 
 # -- Auto-run sphinx-apidoc --------------------------------------------------
 
-# def run_apidoc(_):
-#     from sphinx.ext.apidoc import main
-#     import os
-#     import sys
-#     sys.path.append(os.path.join(os.path.dirname(__file__), '..'))
-#     cur_dir = os.path.abspath(os.path.dirname(__file__))
-#     proj_dir = os.path.dirname(os.path.dirname(cur_dir))
-#     doc_dir = os.path.join(proj_dir, "docs")
-#     output_path = os.path.join(doc_dir, "source", "generated")
-#     module = os.path.join(proj_dir, "pyrekordbox")
-#     exclude = os.path.join(module, "tests")
-#     template_dir = os.path.join(doc_dir, "source", "_templates", "apidoc")
-#     main(["-fMeT", "-o", output_path, module, exclude, "--templatedir", template_dir])
-#
-#
-# def setup(app):
-#     app.connect('builder-inited', run_apidoc)
+
+def run_apidoc(_):
+    from sphinx.ext.apidoc import main
+    import os
+    import sys
+
+    sys.path.append(os.path.join(os.path.dirname(__file__), ".."))
+    cur_dir = os.path.abspath(os.path.dirname(__file__))
+    proj_dir = os.path.dirname(os.path.dirname(cur_dir))
+    doc_dir = os.path.join(proj_dir, "docs")
+    output_path = os.path.join(doc_dir, "source", "generated")
+    module = os.path.join(proj_dir, "pyrekordbox")
+    exclude = os.path.join(module, "tests")
+    template_dir = os.path.join(doc_dir, "source", "_templates", "apidoc")
+    main(["-fMeT", "-o", output_path, module, exclude, "--templatedir", template_dir])
+
+
+def setup(app):
+    app.connect("builder-inited", run_apidoc)