+++ CHANGELOG.md
@@ -1,15 +1,22 @@
 # What's New
 
-<a name="unreleased"></a>
-## [Unreleased]
-
+<a name="0.1.1"></a>
+## [0.1.1] - 2022-19-10
 ### New Features
 - `AnlzFile` now stores the path of the parsed file
 - add `update_content_path` and `update_content_filename` to RB6 database
 
 ### Improvements/Bug Fixes
+- fix bugs in PQTZ/PQT2 tag handler of ANLZ files
+- improve ANLZ file path handling
+- fix bug when reading the pyrekordbox config files
 - remove wrong type hint in ``AbstractAnlzTag``
 - Use path instead of extension as key in `read_anlz_files` output
+
+### Documentation
+- update Quick-Start and change reference labels
+- add initial version of API Reference
+- fix links in ANLZ file documentation
 
 
 <a name="0.1.0"></a>
@@ -134,7 +141,8 @@
 ## 0.0.0 - 2022-10-04
 
 
-[Unreleased]: https://github.com/dylanljones/pyrekordbox/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/dylanljones/pyrekordbox/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/dylanljones/pyrekordbox/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/dylanljones/pyrekordbox/compare/0.0.8...0.1.0
 [0.0.8]: https://github.com/dylanljones/pyrekordbox/compare/0.0.7...0.0.8
 [0.0.7]: https://github.com/dylanljones/pyrekordbox/compare/0.0.6...0.0.7